requirements.txt
discord.py==1.6.0
mcstatus==6.1.2